Move targeted tests into main suites
