+++ docs/requirements_traceability_matrix.md
@@ -8,7 +8,7 @@
 | F-04 | Hybrid DKG persistence | `storage.py`, `storage_backends.py` | `tests/unit/test_duckdb_storage_backend.py`, `tests/unit/test_duckdb_storage_backend_extended.py`, `tests/unit/test_storage_persistence.py`, `tests/unit/test_storage_teardown.py`, `tests/unit/test_storage_utils.py`, `tests/unit/test_storage_backup.py`, `tests/unit/test_storage_persistence_eviction.py`, `tests/integration/test_rdf_persistence.py`, `tests/behavior/features/dkg_persistence.feature` |
 | F-05 | Configurable loops and agent roster via TOML with hot reload | `config.py`, `main.py` | `tests/unit/test_config_reload.py`, `tests/unit/test_config_profiles.py`, `tests/unit/test_config_watcher_cleanup.py`, `tests/unit/test_main_config_commands.py`, `tests/behavior/features/configuration_hot_reload.feature` |
 | F-06 | Schema-validated JSON and human-readable output | `output_format.py`, `models.py` | `tests/unit/test_output_format.py`, `tests/unit/test_template.py`, `tests/unit/test_models_docstrings.py`, `tests/unit/test_llm_docstrings.py`, `tests/behavior/features/output_formatting.feature` |
-| F-07 | Prometheus metrics and monitoring | `orchestration/metrics.py`, `main.py`, `resource_monitor.py`, `monitor/` | `tests/unit/test_metrics.py`, `tests/integration/test_monitor_metrics.py`, `tests/unit/test_monitor_cli.py`, `tests/targeted/test_metrics_extra.py`, `tests/behavior/features/interactive_monitor.feature` |
+| F-07 | Prometheus metrics and monitoring | `orchestration/metrics.py`, `main.py`, `resource_monitor.py`, `monitor/` | `tests/unit/test_metrics.py`, `tests/integration/test_monitor_metrics.py`, `tests/unit/test_monitor_cli.py`, `tests/unit/test_metrics_extra.py`, `tests/behavior/features/interactive_monitor.feature` |
 | F-08 | Interactive query mode | `main.py` | `tests/behavior/features/query_interface.feature` |
 | F-09 | RAM budget enforcement and eviction | `storage.py` | `tests/unit/test_eviction.py`, `tests/unit/test_storage_eviction.py`, `tests/unit/test_storage_ram_usage.py`, `tests/unit/test_storage_persistence_eviction.py`, `tests/behavior/features/storage_search_integration.feature` |
 | F-10 | Vector search in DuckDB | `storage.py`, `storage_backends.py` | `tests/unit/test_vector_search.py`, `tests/integration/test_vector_extension.py`, `tests/integration/test_vector_extension_extended.py`, `tests/behavior/features/vector_extension_handling.feature`, `tests/behavior/features/storage_search_integration.feature` |
