+++ tests/unit/test_metrics_extra.py
@@ -1,4 +1,5 @@
 import types
+
 from autoresearch.orchestration import metrics
