+++ tests/unit/test_output_format_edgecases2.py
